Client: UDPTransport updated, fix UDP package receiving logic, fix UDP sequence validation logic

+++ Client/Src/UDPConnection.cpp
@@ -35,7 +35,7 @@ uint16_t UDPConnection::TXSeqIncrement()
 
 bool UDPConnection::RXSeqCheckIncrement(uint16_t testSeq)
 {
-    if((rxSeq-testSeq)>(testSeq-rxSeq))
+    if(static_cast<uint16_t>(rxSeq-testSeq)>static_cast<uint16_t>(testSeq-rxSeq))
     {
         rxSeq=testSeq;
         return true;
